+++ src/main/java/ru/netology/controller/PostController.java
@@ -9,7 +9,7 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.Reader;
 
-@Controller
+
 public class PostController {
   public static final String APPLICATION_JSON = "application/json";
   private final PostService service;
