+++ src/main/java/ru/netology/repository/PostRepository.java
@@ -1,44 +1,14 @@
 package ru.netology.repository;
-
-import org.springframework.stereotype.Repository;
-import ru.netology.exception.NotFoundException;
 import ru.netology.model.Post;
-
 import java.util.Collection;
 import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicLong;
-// Stub
-@Repository
-public class PostRepository {
 
-    private final ConcurrentHashMap<Long, Post> repositoryCollection = new ConcurrentHashMap<>();
-    private final AtomicLong atomicLong = new AtomicLong();
+public interface PostRepository {
+    Collection<Post> all();
 
-    public final Collection<Post> all() {
-        return repositoryCollection.values();
-    }
+    Optional<Post> getById(long id);
 
-    public final Optional<Post> getById(long id) {
-        return Optional.ofNullable(repositoryCollection.get(id));
-    }
+    Post save(Post post);
 
-    public final Post save(Post post) {
-
-        if (post.getId() == 0) {
-            long id = atomicLong.incrementAndGet();
-            post.setId(id);
-            repositoryCollection.put(id, post);
-        } else if (post.getId() != 0) {
-            repositoryCollection.replace(post.getId(), post);
-            if (!repositoryCollection.contains(post)) {
-                throw new NotFoundException("Нельзя добавить элемент с номером != 0");
-            }
-        }
-        return post;
-    }
-
-    public final void removeById(long id) {
-        repositoryCollection.remove(id);
-    }
+    void removeById(long id);
 }
